+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+- [Minor] Pass webhook id to handlers
+
 ## [5.2.0] - 2022-10-04
 
 - Added support for the `2022-10` API version [#535](https://github.com/Shopify/shopify-api-node/pull/535)